+++ jabgui/src/main/java/org/jabref/migrations/PerformLoadDatabaseMigrations.java
@@ -7,8 +7,6 @@ import org.jabref.logic.importer.ParserResult;
 
 public class PerformLoadDatabaseMigrations {
 
-    // FIXME: Should be called somewhere
-    // Typical call: performLoadDatabaseMigrations(result, importFormatPreferences.bibEntryPreferences().getKeywordSeparator());
     public static void performLoadDatabaseMigrations(ParserResult parserResult,
                                                      Character keywordDelimited) {
         List<PostOpenMigration> postOpenMigrations = Arrays.asList(